requirements.txt
pyserial==3.3
pyOpenSSL==18.0.0
requests==2.20.0
